Don't use threads anymore
I repeat: DON'T USE THREADS ANYMORE! This is a *huge* rewrite...

- spop now uses the Glib event loop

- Spotify events and notifications are handled using Glib constructs too. In
  particular, the Spotify "notify main thread" callback uses a Glib "idle
  source" instead of an ugly semaphore

- clients are handled through Glib IO channels; they must therefore do only non-
  blocking operations

- gthread is still required for the main loop to be thread-safe (since the
  "notify main thread" can be called from another thread). As a matter of fact,
  spop does not work if gthread is not enabled.

This should fix some bugs (stop random crashes?...), and hopefully make the
development easier in the future.

The "idle" command has not yet been re-implemented in this thread-free context.
This will be done soon if I find out how to do it :)