Restore support for the "idle" command
Clients that issued the "idle" command have their IO channel stored in a GList.
When a notification occurs, the current status is sent to each channel in this
list.

We now explicitly tell which queue_* function can send a notification, so that a
queue_* function that calls another queue_* function will not trigger two
notifications.