New plugin: awesome
This plugin is designed to work with Awesome, a great (tiling) window manager.
It uses D-Bus to communicate with it and update a widget so that it displays the
current status of spop.

When used together, the "awesome" and the "notify" plugins do exactly the same
as the former awesome/spop.py script (except that they use less resources and
are more reliable). For this reason, this script has now been deleted as well.