Make commands independent from the interface
Before this, command_context had a GIOChannel member, and command_end
explicitely called interface_finalize(). This made it impossible to call
commands asynchronously from a plugin.

Now, command_context has a pointer to a finalize function instead. This is more
generic and should allow to do nice things in a near future.