New audio plugin: SoX
This is still work in progress and very experimental, but seems to work for me.

This new plugin uses libsox for audio output. Compared to libao, it adds support
for more output devices (including several platforms probably not supported by
libspotify anyway...), for saving audio output to a file (which is probably not
permitted by Spotify, so please don't do it), and most importantly to using
effects and filter on the audio output: normaliser, equaliser, reverb, etc.

See `man soxeffects' for more details about the effects you can use.