New plugin: web
Adds an HTTP JSON API that can be used to run almost all "normal"
commands: ls, play, status, etc. The idle command is not supported
yet (but will be very soon), and special commands such as bye and quit
won't be supported.

The long-term goal of this plugin is to provide a full web interface to
spop (designed for smartphones). In the future it will therefore serve
a web application made of HTML/CSS/JS files that will make use of this
new API.